refactor: compose running view components
